% Auto-generated: do not edit by hand
\name{''MobileDropdown}

\alias{''MobileDropdown}

\title{MobileDropdown component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''MobileDropdown(id=NULL, className=NULL, clearable=NULL, notfoundMsg=NULL,
options=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{className}{Character. }

\item{clearable}{Logical. }

\item{notfoundMsg}{Character. }

\item{options}{Unnamed list. }

\item{value}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

